:art: refactor(paginate): Remove redundant `Links` and `Meta` pydantic types
Originally, I had just some named tuples for links and meta pagination information. Once we decided to move away from that, it got overlooked. Also, I held on the thought that I might need a mapping from the data and the pydantic response. It was never needed - YAGNI!

GTC-1865
