+++ tests_v2/unit/app/utils/paginate/test_pagination_links_info.py
@@ -2,7 +2,8 @@ from unittest.mock import AsyncMock
 
 import pytest
 
-from app.utils.paginate import PaginationLinks, paginate_collection
+from app.models.pydantic.responses import PaginationLinks
+from app.utils.paginate import paginate_collection
 
 DONT_CARE: int = 1
